pyproject.toml
[tool.poetry]
name = "prodexa"
version = "0.1.0"
description = "A application to automatize Monalise's exams creation"
authors = ["André Carvalho <alac1984@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^0.8.11"
pillow = "^9.4.0"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
pytest = "^7.2.2"
pre-commit = "^3.2.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 95
